handling current location, filtering mobs, half map size variable, drawPlayerSelf() function, uint64_t id instead of uint32_t, finally corect change location handling, much more mobs, sortMobDescriptions function for printing sorted mob list

+++ packetAnalyze/HealthUpdate.h
@@ -3,12 +3,12 @@
 
 struct HealthUpdate
 {
-	uint32_t _id;
+	uint64_t _id;
 	uint32_t _healthChange;
 	uint32_t _health;
 
 	HealthUpdate(NetworkCommand& rawHealthUpdate);
-	HealthUpdate(uint32_t id, uint32_t healthChange, uint32_t health);
+	HealthUpdate(uint64_t id, uint32_t healthChange, uint32_t health);
 
 };
